Move shared functionality to base offline maps fragment
