--- opensrp-reveal/src/main/java/org/smartregister/reveal/contract/OfflineMapsContract.java
@@ -1,11 +1,0 @@
-package org.smartregister.reveal.contract;
-
-public class OfflineMapsContract {
-
-    interface Presenter extends BaseContract.BasePresenter {
-
-    }
-
-    public interface View {
-    }
-}
