+++ opensrp-reveal/src/main/java/org/smartregister/reveal/contract/OfflineMapsFragmentContract.java
@@ -1,0 +1,13 @@
+package org.smartregister.reveal.contract;
+
+public interface OfflineMapsFragmentContract {
+
+    interface View {
+
+        void displayToast(String message);
+
+        void displayError(int title, String message);
+
+    }
+
+}
