+++ opensrp-reveal/src/main/java/org/smartregister/reveal/fragment/AvailableOfflineMapsFragment.java
@@ -1,22 +1,13 @@
 package org.smartregister.reveal.fragment;
 
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.widget.RecyclerView;
-import android.text.TextUtils;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.Toast;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -41,10 +32,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import io.ona.kujaku.helpers.OfflineServiceHelper;
-import io.ona.kujaku.services.MapboxOfflineDownloaderService;
-import io.ona.kujaku.utils.Constants;
 
-public class AvailableOfflineMapsFragment extends Fragment implements AvailableOfflineMapsContract.View {
+public class AvailableOfflineMapsFragment extends BaseOfflineMapsFragment implements AvailableOfflineMapsContract.View {
 
     private RecyclerView offlineMapRecyclerView;
 
@@ -53,10 +42,6 @@ public class AvailableOfflineMapsFragment extends Fragment implements AvailableO
     private AvailableOfflineMapsPresenter presenter;
 
     private List<OfflineMapModel> offlineMapModelList = new ArrayList<>();
-
-    private String currentMapDownload;
-
-    private MapDownloadReceiver mapDownloadReceiver = new MapDownloadReceiver();
 
     private List<Location> operationalAreasToDownload = new ArrayList<>();
 
@@ -75,19 +60,6 @@ public class AvailableOfflineMapsFragment extends Fragment implements AvailableO
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         presenter = new AvailableOfflineMapsPresenter(this);
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        IntentFilter intentFilter = new IntentFilter(Constants.INTENT_ACTION_MAP_DOWNLOAD_SERVICE_STATUS_UPDATES);
-        LocalBroadcastManager.getInstance(getActivity().getApplicationContext()).registerReceiver(mapDownloadReceiver, intentFilter);
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        LocalBroadcastManager.getInstance(getActivity().getApplicationContext()).unregisterReceiver(mapDownloadReceiver);
     }
 
     @Nullable
@@ -135,16 +107,6 @@ public class AvailableOfflineMapsFragment extends Fragment implements AvailableO
             adapter.setOfflineMapModels(offlineMapModelList);
             this.offlineMapModelList = offlineMapModelList;
         }
-    }
-
-    @Override
-    public void displayToast(String message) {
-        Toast.makeText(getContext(), message, Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void displayError(int title, String message) {
-        Toast.makeText(getContext(), message, Toast.LENGTH_LONG).show();
     }
 
     @Override
@@ -241,87 +203,18 @@ public class AvailableOfflineMapsFragment extends Fragment implements AvailableO
         );
     }
 
-    private class MapDownloadReceiver extends BroadcastReceiver {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            Bundle bundle = intent.getExtras();
-            if (bundle != null) {
-                Log.i("KUJAKU SAMPLE APP TAG", intent.getExtras().toString());
-                if (bundle.containsKey(MapboxOfflineDownloaderService.KEY_RESULT_STATUS)
-                        && bundle.containsKey(MapboxOfflineDownloaderService.KEY_RESULT_MESSAGE)
-                        && bundle.containsKey(MapboxOfflineDownloaderService.KEY_RESULTS_PARENT_ACTION)
-                        && bundle.containsKey(Constants.PARCELABLE_KEY_MAP_UNIQUE_NAME)) {
-
-                    String mapUniqueName = bundle.getString(Constants.PARCELABLE_KEY_MAP_UNIQUE_NAME);
-                    String resultStatus = bundle.getString(MapboxOfflineDownloaderService.KEY_RESULT_STATUS);
-                    MapboxOfflineDownloaderService.SERVICE_ACTION serviceAction = (MapboxOfflineDownloaderService.SERVICE_ACTION) bundle.get(MapboxOfflineDownloaderService.KEY_RESULTS_PARENT_ACTION);
-
-                    String message = bundle.getString(MapboxOfflineDownloaderService.KEY_RESULT_MESSAGE);
-
-                    if (MapboxOfflineDownloaderService.SERVICE_ACTION_RESULT.FAILED.name().equals(resultStatus)) {
-                        if (!TextUtils.isEmpty(message)) {
-                            if (!message.contains("MapBox Tile Count limit exceeded")) {
-                                //showInfoNotification("Error occurred " + mapUniqueName + ":" + serviceAction.name(), message);
-
-                                displayError(R.id.download_map,  message);
-                            }
-                        }
-
-                        if (serviceAction == MapboxOfflineDownloaderService.SERVICE_ACTION.DELETE_MAP && !TextUtils.isEmpty(message)) {
-                           displayError(R.id.download_map, message);
-                        }
-                                    /*
-                                    (FACT) This is an error update from the service. If this is not
-                                    a DELETE_MAP action and the update is about the map that we expect
-                                    to be currently downloading, held by currentMapDownload variable, then we
-                                    need to disable the STOP MAP DOWNLOAD since the download has already been
-                                    stopped after the error. If we left this as true, then we would be misleading
-                                    the user that they can stop a non-existent download.
-                                     */
-                        else if (!TextUtils.isEmpty(mapUniqueName) && mapUniqueName.equals(currentMapDownload)) {
-                           // setCanStopMapDownload(false);
-                        }
-                    } else {
-                        // We should disable the stop offline download button if it was stopped successfully
-                        if (serviceAction == MapboxOfflineDownloaderService.SERVICE_ACTION.STOP_CURRENT_DOWNLOAD) {
-                            currentMapDownload = null;
-                           // setCanStopMapDownload(false);
-                            displayToast("Download stopped");
-                        } else {
-                            if (!TextUtils.isEmpty(message)) {
-                                // This is a download progress message
-                                if (isValidDouble(message)) {
-                                    if (Double.valueOf(message) == 100d) {
-                                        currentMapDownload = null;
-                                        displayToast("Download finished successfuly");
-                                        presenter.onDownloadComplete(mapUniqueName);
-                                       // setCanStopMapDownload(false);
-                                    } else {
-                                       // setCanStopMapDownload(true);
-                                        displayToast("Download map for " + mapUniqueName + " in progress at " + Double.valueOf(message));
-                                        presenter.onDownloadStarted(mapUniqueName);
-                                    }
-                                } else {
-                                    displayToast(message);
-                                }
-                            }
-                        }
-                    }
-                }
-            } else {
-                Log.i("Reveal APP TAG", "Broadcast message has null Extras");
-            }
-
-        }
-    }
-
-    private boolean isValidDouble(String doubleString) {
-        String doubleRegex = "[+-]{0,1}[0-9]*.{0,1}[0-9]*";
-        return (!doubleString.isEmpty() && doubleString.matches(doubleRegex));
-    }
-
     public void setOfflineMapDownloadCallback(OfflineMapDownloadCallback callBack) {
         this.callback = callBack;
+    }
+
+    @Override
+    protected void downloadCompleted(String mapUniqueName) {
+        presenter.onDownloadComplete(mapUniqueName);
+    }
+
+    @Override
+    protected void downloadStarted(String mapUniqueName) {
+        presenter.onDownloadStarted(mapUniqueName);
     }
 
 }
