+++ opensrp-reveal/src/main/java/org/smartregister/reveal/view/OfflineMapsActivity.java
@@ -9,12 +9,11 @@ import android.support.v7.app.AppCompatActivity;
 import org.smartregister.family.adapter.ViewPagerAdapter;
 import org.smartregister.reveal.R;
 import org.smartregister.reveal.contract.OfflineMapDownloadCallback;
-import org.smartregister.reveal.contract.OfflineMapsContract;
 import org.smartregister.reveal.fragment.AvailableOfflineMapsFragment;
 import org.smartregister.reveal.fragment.DownloadedOfflineMapsFragment;
 import org.smartregister.reveal.model.OfflineMapModel;
 
-public class OfflineMapsActivity extends AppCompatActivity implements OfflineMapsContract.View, OfflineMapDownloadCallback {
+public class OfflineMapsActivity extends AppCompatActivity implements OfflineMapDownloadCallback {
 
     private ViewPagerAdapter adapter;
